Moved Sentiment to root directory
