+++ src/predict/music_creator/sentiment_to_melodies.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import Optional, Type
 from music21 import instrument, articulations
 import random
-from sentiment_classifier.sentiment import Sentiment
+from src.sentiment import Sentiment
 
 
 @dataclass
