+++ src/predict/sentiment_classifier/sentiment_classifier.py
@@ -1,5 +1,5 @@
 import numpy as np
-from sentiment_classifier.sentiment import Sentiment
+from src.sentiment import Sentiment
 from tensorflow.keras.preprocessing.sequence import pad_sequences
 from tensorflow.keras.preprocessing.text import Tokenizer
 from tensorflow.keras.models import Model
